Working on all component endpoints

+++ db/characters/delete_character.sql
@@ -1,3 +1,7 @@
 DELETE FROM characters
-WHERE char_id = $1;
-SELECT * FROM characters;
+WHERE character_id = $2
+AND user_id = $1;
+SELECT * 
+FROM characters
+WHERE user_id = $1
+ORDER BY character_id;
