+++ db/locations/delete_location.sql
@@ -1,3 +1,7 @@
 DELETE FROM locations
-WHERE loc_id = $1;
-SELECT * FROM locations;
+WHERE location_id = $2
+AND user_id = $1;
+SELECT *
+FROM locations
+WHERE user_id = $1
+ORDER BY location_id;
