+++ db/progress/delete_progress.sql
@@ -1,3 +1,7 @@
 DELETE FROM progress
-WHERE log_id = $1;
-SELECT * FROM progress;
+WHERE progress_id = $2
+AND user_id = $1;
+SELECT *
+FROM progress
+WHERE user_id = $1
+ORDER BY progress_id;
